cros: vboot: Fix warning in cmd_vboot_test.c
Addresses a compilation warning when building Sandbox U-Boot.

BRANCH=none
BUG=chromium-os:32603
TEST=Prevents warning about cast from pointer to int of different size
Reviewed-on: https://gerrit.chromium.org/gerrit/30611

+++ cros/cmd/cmd_vboot_test.c
@@ -217,7 +217,7 @@ static void show_ec_bin(const char *name, struct fmap_entry *entry)
 	printf("EC binary %s at %#x, length %#x\n", name, entry->offset,
 	       entry->length);
 	if (entry->length) {
-		data = (u32 *)(CONFIG_SYS_TEXT_BASE + entry->offset);
+		data = (u32 *)(uintptr_t)(CONFIG_SYS_TEXT_BASE + entry->offset);
 		printf("%p: ", data);
 		for (i = 0; i < 4; i++)
 			printf("%08x ", data[i]);
